pyproject.toml
[project]
name = "skincancervit"
version = "0.1.0"
description = "A multimodal Vision Transformer model for skin cancer diagnosis, combining image and tabular data."
readme = "README.md"
requires-python = ">=3.10.17"
dependencies = [
    "datasets>=4.0.0",
    "evaluate>=0.4.5",
    "grad-cam",
    "gradio>=5.38.0",
    "imageio>=2.37.0",
    "pillow>=11.3.0",
    "safetensors>=0.5.3",
    "scikit-learn>=1.7.0",
    "torch>=2.7.1",
    "transformers[torch]>=4.53.2",
]

[tool.ruff.lint]

# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`)  codes by default.
select = ["E4", "E7", "E9", "F"]
ignore = []

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

[tool.ruff.format]

# Double quotes for strings.
quote-style = "double"

# Iindent with spaces.
indent-style = "space"

# Respect magic trailing commas.
skip-magic-trailing-comma = false

# Automatically detect line ending.
line-ending = "auto"

# Disable auto-formatting of code examples in docstrings.
docstring-code-format = false

# Set the line length limit used when formatting code snippets in
# docstrings.
#
# This only has an effect when the `docstring-code-format` setting is
# enabled.
docstring-code-line-length = "dynamic"

[tool.ruff.lint.pydocstyle]
convention = "google"  # Accepts: "google", "numpy", or "pep257".

[build-system]
requires = ["setuptools>=61.0", "wheel", "build"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["skincancer_vit"]
exclude = ["results*"]

[tool.uv.sources]
grad-cam = { git = "https://github.com/jacobgil/pytorch-grad-cam.git" }
